+++ vscp/events/vscp_information.h
@@ -405,6 +405,7 @@ extern BOOL vscp_information_sendGoodbyeEvent(uint8_t zone, uint8_t subZone);
 /**
  * A node indicates that a stop event occurred. This can for example be a motor stopping.
  *
+ * @param[in] reserved Reserved for any kind of information
  * @param[in] zone Zone for which event applies to (0-255). 255 is all zones.
  * @param[in] subZone Sub-zone for which event applies to (0-255). 255 is all sub-zones.
  * @return Status
@@ -412,7 +413,7 @@ extern BOOL vscp_information_sendGoodbyeEvent(uint8_t zone, uint8_t subZone);
  * @retval TRUE  Event successul sent
  *
  */
-extern BOOL vscp_information_sendStopEvent(uint8_t zone, uint8_t subZone);
+extern BOOL vscp_information_sendStopEvent(uint8_t reserved, uint8_t zone, uint8_t subZone);
 
 /**
  * A node indicates that a start event occurred. This can be a motor starting.
@@ -967,6 +968,58 @@ extern BOOL vscp_information_sendAstronomicalSunsetTwilightTimeEvent(uint8_t use
  *
  */
 extern BOOL vscp_information_sendCalculatedNoonEvent(uint8_t zone, uint8_t subZone);
+
+/**
+ * A node indicates that something is driving up, e. g. shutter is driving up.
+ *
+ * @param[in] index Index
+ * @param[in] zone Zone for which event applies to (0-255). 255 is all zones.
+ * @param[in] subZone Sub-zone for which event applies to (0-255). 255 is all sub-zones.
+ * @return Status
+ * @retval FALSE Failed to send the event
+ * @retval TRUE  Event successul sent
+ *
+ */
+extern BOOL vscp_information_sendUpEvent(uint8_t index, uint8_t zone, uint8_t subZone);
+
+/**
+ * A node indicates that something is driving down, e. g. shutter is driving down.
+ *
+ * @param[in] index Index
+ * @param[in] zone Zone for which event applies to (0-255). 255 is all zones.
+ * @param[in] subZone Sub-zone for which event applies to (0-255). 255 is all sub-zones.
+ * @return Status
+ * @retval FALSE Failed to send the event
+ * @retval TRUE  Event successul sent
+ *
+ */
+extern BOOL vscp_information_sendDownEvent(uint8_t index, uint8_t zone, uint8_t subZone);
+
+/**
+ * A node indicates that something is at the top, e. g. shutter is at the top.
+ *
+ * @param[in] index Index
+ * @param[in] zone Zone for which event applies to (0-255). 255 is all zones.
+ * @param[in] subZone Sub-zone for which event applies to (0-255). 255 is all sub-zones.
+ * @return Status
+ * @retval FALSE Failed to send the event
+ * @retval TRUE  Event successul sent
+ *
+ */
+extern BOOL vscp_information_sendTopEvent(uint8_t index, uint8_t zone, uint8_t subZone);
+
+/**
+ * A node indicates that something is at the bottom, e. g. shutter is at the bottom.
+ *
+ * @param[in] index Index
+ * @param[in] zone Zone for which event applies to (0-255). 255 is all zones.
+ * @param[in] subZone Sub-zone for which event applies to (0-255). 255 is all sub-zones.
+ * @return Status
+ * @retval FALSE Failed to send the event
+ * @retval TRUE  Event successul sent
+ *
+ */
+extern BOOL vscp_information_sendBottomEvent(uint8_t index, uint8_t zone, uint8_t subZone);
 
 #endif /* __VSCP_INFORMATION_H__ */